requirements.txt
aiofiles~=0.7
aiosqlite~=0.17
apscheduler<4,>=3.8
feedparser<7,>=6.0
hikari[speedups]==2.0.0.dev104
hikari-lightbulb<3,>=2.0.4
piston-rspy~=0.2
psutil<6,>=5.8
pygithub<2,>=1.55
pygount<2,>=1.2
python-dotenv~=0.19
uvloop~=0.16; os_name != "nt"
